fix burger number and other mistakes
