+++ src/components/burger-ingredients/ingredient/ingredient-types.ts
@@ -17,4 +17,5 @@ export type TIngredient = {
     __v: number;
     index?: number;
     ingredientCount: number;
+    ingredientOrderCount: number;
 };
